+++ nautobot_circuit_maintenance/tests/test_handler.py
@@ -1,4 +1,5 @@
 """Tests for Handle Notifications methods."""
+import uuid
 from unittest.mock import Mock, patch
 from datetime import datetime, timezone
 from email.message import EmailMessage
@@ -61,13 +62,13 @@ END:VCALENDAR
     email_message = EmailMessage()
     email_message["From"] = "Sender <sender@example.com>"
     email_message["Date"] = format_datetime(notification_data["stamp"])
-    email_message["Subject"] = "Test subject"
+    email_message["Subject"] = "Test subject " + "1234567890" * 20
     email_message["Content-Type"] = "text/calendar"
     email_message.set_payload(template.render(obj=notification_data).encode("utf-8"))
 
     return MaintenanceNotification(
         msg_id=b"12345",
-        subject="Test subject",
+        subject="Test subject " + "1234567890" * 20,
         sender="sender@example.com",
         source=source,
         raw_payload=email_message.as_bytes(),
@@ -164,6 +165,20 @@ class TestHandleNotificationsJob(TestCase):  # pylint: disable=too-many-public-m
             mock_tag_message.assert_any_call(self.job, test_notification.msg_id, "parsed")
             mock_tag_message.assert_any_call(self.job, test_notification.msg_id, "unknown-cids")
             self.job.log_debug.assert_called_with("1 notifications processed.")
+
+            # Do some checking of string representation length for potential change-logging issues
+            # (ObjectChange.object_repr field has a limit of 200 characters)
+            for model in (
+                RawNotification,
+                ParsedNotification,
+                CircuitMaintenance,
+                CircuitImpact,
+                Note,
+            ):
+                objectchange = model.objects.first().to_objectchange("create")
+                objectchange.request_id = uuid.uuid4()
+                objectchange.save()
+                self.assertLessEqual(len(str(model.objects.first())), 200)
 
     def test_run_no_notifications(self):
         """Test when a there are no notifications."""
@@ -335,7 +350,7 @@ class TestHandleNotificationsJob(TestCase):  # pylint: disable=too-many-public-m
         test_notification = generate_email_notification(notification_data, self.source)
         provider = Provider.objects.get(slug=test_notification.provider_type)
         RawNotification.objects.get_or_create(
-            subject=test_notification.subject,
+            subject=test_notification.subject[:MAX_NOTIFICATION_SUBJECT_LENGTH],
             provider=provider,
             raw=test_notification.raw_payload,
             sender=test_notification.sender,
@@ -363,7 +378,7 @@ class TestHandleNotificationsJob(TestCase):  # pylint: disable=too-many-public-m
         test_notification = generate_email_notification(notification_data, self.source)
         provider = Provider.objects.get(slug=test_notification.provider_type)
         RawNotification.objects.get_or_create(
-            subject=test_notification.subject,
+            subject=test_notification.subject[:MAX_NOTIFICATION_SUBJECT_LENGTH],
             provider=provider,
             raw=test_notification.raw_payload,
             sender=test_notification.sender,
@@ -386,6 +401,37 @@ class TestHandleNotificationsJob(TestCase):  # pylint: disable=too-many-public-m
         self.assertEqual(0, len(CircuitImpact.objects.all()))
         self.assertEqual(1, len(Note.objects.all()))
         mock_tag_message.assert_called_with(self.job, test_notification.msg_id, "unknown-cids")
+
+    def test_create_circuit_maintenance_unknown_status(self):
+        """Test create_circuit_maintenance with an unknown status."""
+        notification_data = get_base_notification_data()
+        test_notification = generate_email_notification(notification_data, self.source)
+        provider = Provider.objects.get(slug=test_notification.provider_type)
+        RawNotification.objects.get_or_create(
+            subject=test_notification.subject[:MAX_NOTIFICATION_SUBJECT_LENGTH],
+            provider=provider,
+            raw=test_notification.raw_payload,
+            sender=test_notification.sender,
+            source=self.notification_source,
+            stamp=datetime.now(timezone.utc),
+        )
+        parser_provider = init_provider(provider_type=test_notification.provider_type)
+        data_to_process = NotificationData.init_from_email_bytes(test_notification.raw_payload)
+        parsed_maintenance = parser_provider.get_maintenances(data_to_process)[0]
+        parsed_maintenance.status = "No idea!"
+        create_circuit_maintenance(
+            self.job,
+            test_notification,
+            f"{provider.slug}-{parsed_maintenance.maintenance_id}",
+            parsed_maintenance,
+            provider,
+        )
+
+        self.assertEqual(1, len(CircuitMaintenance.objects.all()))
+        self.assertEqual(2, len(CircuitImpact.objects.all()))
+        self.assertEqual(0, len(Note.objects.all()))
+
+        self.assertEqual("UNKNOWN", CircuitMaintenance.objects.first().status)
 
     def test_update_circuit_maintenance(self):
         """Test update_circuit_maintenance."""
@@ -450,6 +496,31 @@ class TestHandleNotificationsJob(TestCase):  # pylint: disable=too-many-public-m
         self.assertEqual(circuit_maintenance_entry.status, "COMPLETED")
         # Verify that both parsed notifications are linked to the CircuitMaintenance for future reference
         self.assertEqual(len(circuit_maintenance_entry.parsednotification_set.all()), 2)
+
+    def test_update_circuit_maintenance_status_no_change(self):
+        """Test update_circuit_maintenance with a "NO-CHANGE" status value."""
+        notification_data = get_base_notification_data()
+        test_notification = generate_email_notification(notification_data, self.source)
+        provider = Provider.objects.get(slug=test_notification.provider_type)
+        with patch(
+            "nautobot_circuit_maintenance.handle_notifications.handler.get_notifications"
+        ) as mock_get_notifications:
+            mock_get_notifications.return_value = [test_notification]
+            self.job.run(commit=True)
+
+        # Adding changes
+        parser_provider = init_provider(provider_type=test_notification.provider_type)
+        data_to_process = NotificationData.init_from_email_bytes(test_notification.raw_payload)
+        parsed_maintenance = parser_provider.get_maintenances(data_to_process)[0]
+        parsed_maintenance.status = "NO-CHANGE"
+        maintenance_id = f"{provider.slug}-{parsed_maintenance.maintenance_id}"
+        circuit_maintenance_entry = CircuitMaintenance.objects.get(name=maintenance_id)
+
+        update_circuit_maintenance(self.job, test_notification, circuit_maintenance_entry, parsed_maintenance, provider)
+
+        circuit_maintenance_entry = CircuitMaintenance.objects.get(name=maintenance_id)
+        # Status should not be changed:
+        self.assertEqual("CONFIRMED", circuit_maintenance_entry.status)
 
     def test_create_or_update_circuit_maintenance_truncated_fields(self):
         """Test create_or_update_circuit_maintenance with long fields that must be truncated to fit the DB."""
@@ -523,7 +594,7 @@ class TestHandleNotificationsJob(TestCase):  # pylint: disable=too-many-public-m
         test_notification = generate_email_notification(notification_data, self.source)
         provider = Provider.objects.get(slug=test_notification.provider_type)
         RawNotification.objects.get_or_create(
-            subject=test_notification.subject,
+            subject=test_notification.subject[:MAX_NOTIFICATION_SUBJECT_LENGTH],
             provider=provider,
             raw=test_notification.raw_payload,
             sender=test_notification.sender,