+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## v0.4.3
+
+### Fixed
+
+- #172 - Fix potential `DataError: value too long for type` when handling especially long email subjects
+
+## v0.4.2
+
+### Fixed
+
+- #162 - Fix inability to refresh OAuth authentication before token expires
+- #166 - Handle `NO-CHANGE` status from parser (networktocode/circuit-maintenance-parser#125), handle unknown statuses from parser, fix potential error during `nautobot-server post_migrate` signal handling.
+- #168 - Discard cached OAuth token if refreshing it fails.
+
+## v0.4.1 - 2021-11-29
+
+### Fixed
+
+- #158: Removed `utf-8` decoding for `IMAP` source to support for `ascii` encoded emails.
+
 ## v0.4.0 - 2021-11-11
 
 ### Added